Tests/LibWeb/Text/expected/css/additive-symbols-descriptor-calculated-values.txt

@counter-style valid { additive-symbols: calc(1 + sign(1em - 1px)) "A", sign(1em - 1px) "B"; }
@counter-style invalid { }
@counter-style valid { additive-symbols: 1 "E", calc(-1) "F"; }
